+++ TouchDetector/MotionRecorder.h
@@ -5,7 +5,7 @@ struct Motion
 {
 	int currentTimestamp;
 	int id;
-	std::vector<std::pair<short, short>> points;
+	std::vector<std::pair<float, float>> points;
 };
 
 class MotionRecorder
@@ -17,7 +17,7 @@ public:
 	MotionRecorder();
 	~MotionRecorder();
 
-	void track(std::vector<std::pair<short, short>> points);
+	void track(std::vector<std::pair<float, float>> points);
 	void onTouch(const Motion& motion);
 	void onRelease(const Motion& motion);
 	const std::vector<Motion>& motions() const;
